Delete index.php and Test Branch Merged Note
I merged the test branch which brought the index.php file into the
Master branch. This is a way to work on new features for a program and
then implementing the new feature once it is complete. Committing to
the Master branch should be for bug fixes while making a new branch,
committing, and then merging is for larger feature roll outs.